change_apply_location_into_auth
