+++ lib/core/routes/app_router.dart
@@ -3,16 +3,15 @@ import 'package:flowery_rider/core/di/injectable.dart';
 import 'package:flowery_rider/application_approved_page.dart';
 import 'package:flowery_rider/core/routes/routes.dart';
 import 'package:flowery_rider/features/Home_layout/home_layout.dart';
-import 'package:flowery_rider/features/apply/presentation/cubit/auth_cubit.dart';
-import 'package:flowery_rider/features/apply/presentation/pages/apply_page.dart';
-import 'package:flowery_rider/features/apply/presentation/pages/apply_success_page.dart';
-import 'package:flowery_rider/features/auth/presentation/pages/login_screen/login_screen.dart';
-import 'package:flowery_rider/features/auth/presentation/pages/login_screen/login_cubit.dart';
-import 'package:flowery_rider/features/auth/presentation/pages/onboarding/on_boarding_screen.dart';
+import 'package:flowery_rider/features/auth/presentation/apply/cubit/auth_cubit.dart';
+import 'package:flowery_rider/features/auth/presentation/apply/pages/apply_page.dart';
+import 'package:flowery_rider/features/auth/presentation/apply/pages/apply_success_page.dart';
+import 'package:flowery_rider/features/auth/presentation/login_screen/login_screen.dart';
+import 'package:flowery_rider/features/auth/presentation/login_screen/login_cubit.dart';
 import 'package:flowery_rider/features/forget_password/presentation/pages/reset_password_page.dart';
+import 'package:flowery_rider/features/onboarding/on_boarding_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
 import '../../features/forget_password/presentation/cubit/forget_password_cubit.dart';
 import '../../features/forget_password/presentation/pages/forget_password_page.dart';
 import '../../features/forget_password/presentation/pages/pin_code_page.dart';
