+++ lib/features/auth/domain/entities/apply_entity.dart
