+++ lib/features/auth/presentation/apply/cubit/auth_state.dart
