+++ lib/features/auth/presentation/apply/pages/apply_page.dart
@@ -1,4 +1,4 @@
-// features/apply/presentation/pages/apply_page.dart
+// features/auth/presentation/apply/pages/apply_page.dart
 import 'dart:convert';
 import 'dart:io';
 import 'package:flowery_rider/core/base/base_state.dart';
@@ -7,14 +7,14 @@ import 'package:flowery_rider/core/theme/app_colors.dart';
 import 'package:flowery_rider/core/theme/app_styles.dart';
 import 'package:flowery_rider/core/utils/validator.dart';
 import 'package:flowery_rider/core/widget/dialog_utils.dart';
-import 'package:flowery_rider/features/apply/data/model/apply/country_data.dart';
-import 'package:flowery_rider/features/apply/data/model/apply/country_model.dart';
-import 'package:flowery_rider/features/apply/domain/entities/apply_entity.dart';
-import 'package:flowery_rider/features/apply/presentation/cubit/auth_cubit.dart';
-import 'package:flowery_rider/features/apply/presentation/widgets/apply_widgets/form_section.dart';
-import 'package:flowery_rider/features/apply/presentation/widgets/apply_widgets/gender_selection.dart';
-import 'package:flowery_rider/features/apply/presentation/widgets/apply_widgets/searchable_dropdown_field.dart';
-import 'package:flowery_rider/features/apply/presentation/widgets/apply_widgets/upload_field.dart';
+import 'package:flowery_rider/features/auth/data/models/apply/country_data.dart';
+import 'package:flowery_rider/features/auth/data/models/apply/country_model.dart';
+import 'package:flowery_rider/features/auth/domain/entities/apply_entity.dart';
+import 'package:flowery_rider/features/auth/presentation/apply/cubit/auth_cubit.dart';
+import 'package:flowery_rider/features/auth/presentation/apply/widgets/form_section.dart';
+import 'package:flowery_rider/features/auth/presentation/apply/widgets/gender_selection.dart';
+import 'package:flowery_rider/features/auth/presentation/apply/widgets/searchable_dropdown_field.dart';
+import 'package:flowery_rider/features/auth/presentation/apply/widgets/upload_field.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
