+++ lib/features/auth/presentation/apply/pages/apply_success_page.dart
