+++ lib/features/auth/presentation/apply/widgets/vehicle_type_dropdown.dart
