+++ lib/features/auth/presentation/login_screen/login_screen.dart
@@ -1,12 +1,13 @@
+// features/auth/presentation/login_screen/login_screen.dart
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flowery_rider/core/base/base_state.dart';
 import 'package:flowery_rider/core/di/injectable.dart';
 import 'package:flowery_rider/core/routes/routes.dart';
 import 'package:flowery_rider/core/theme/app_colors.dart';
 import 'package:flowery_rider/core/widget/dialog_utils.dart';
-import 'package:flowery_rider/features/auth/presentation/pages/login_screen/login_cubit.dart';
-import 'package:flowery_rider/features/auth/presentation/pages/login_screen/login_states.dart';
-import 'package:flowery_rider/features/auth/presentation/pages/login_screen/login_widgets/login_form.dart';
+import 'package:flowery_rider/features/auth/presentation/login_screen/login_cubit.dart';
+import 'package:flowery_rider/features/auth/presentation/login_screen/login_states.dart';
+import 'package:flowery_rider/features/auth/presentation/login_screen/login_widgets/login_form.dart';
 import 'package:flowery_rider/generated/locale_keys.g.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
